Rename the file with rebase-helper methods, add docstring
